+++ src/lib/openjp2/segmented_stream.h
@@ -84,7 +84,7 @@ struct opj_seg_buf_t {
 	opj_seg_buf_t();
 	~opj_seg_buf_t();
     size_t data_len;	/* total length of all segments*/
-    int32_t cur_seg_id;	/* current index into segments vector */
+    size_t cur_seg_id;	/* current index into segments vector */
 	std::vector<opj_buf_t*> segments;
 };
